# docs/survey-tools-kyc.openapi.yaml
openapi: 3.0.3
info:
  title: SurveyTools KYC & Researcher Profile API
  version: 1.0.0
  description: >
    KYC and Researcher Profile endpoints for SurveyTools.
    Includes unified registration, progressive profiling (Layers 1-5), and OTP verification.
servers:
  - url: "{{base_url}}"
    description: Configured base URL (set per environment in Postman)
paths:
...
  /reset-password:
    post:
      summary: Reset password using OTP
      description: Verifies the 6-digit OTP and sets a new password.
      operationId: resetPassword
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResetPasswordBody'
      responses:
        '200':
          description: Password has been reset successfully
        '400':
          description: Validation error or invalid OTP
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT issued by the existing SurveyTools auth flow
  schemas:
...
    ResetPasswordBody:
      type: object
      required: [email, code, newPassword, confirmPassword]
      properties:
        email: { type: string, format: email }
        code: { type: string, pattern: '^\d{6}$' }
        newPassword: { type: string, minLength: 8 }
        confirmPassword: { type: string }
...